exam_django/requirementsv2.txt
aiohttp==3.9.3
aiosignal==1.3.1
aiosmtplib==1.1.6
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.3
celery[redis]
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.13
django-cors-headers==4.3.1
django-rest-passwordreset==1.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Django==5.0.2
drf-yasg==1.21.7
exceptiongroup==1.2.0
executing==2.0.1
extra-streamlit-components==0.1.71
faiss-cpu==1.8.0
filelock==3.13.1
flake8==7.0.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio-status==1.62.1
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
httplib2==0.22.0
idna==3.6
inflection==0.5.1
ipykernel==6.29.4
ipython==8.23.0
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema-specifications==2023.12.1
jsonschema==4.21.1
jupyter_client==8.6.1
jupyter_core==5.7.2
langchain-community==0.0.31
langchain-core==0.1.40
langchain-google-genai==1.0.1
langchain-text-splitters==0.0.1
langchain==0.1.14
langsmith==0.1.41
Markdown-it-py==3.0.0
marshmallow==3.21.1
mccabe==0.7.0
git+https://github.com/azure-samples/ms-identity-python-samples-common
mdurl==0.1.2
multidict==6.0.5
msal==1.27.0
mysqlclient==2.2.4
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
orjson==3.10.0
packaging==23.2
pandas==2.2.1
parso==0.8.4
phonenumbers==8.12.48
pillow==10.3.0
pkce==1.0.3
platformdirs==4.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.9
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1-modules==0.4.0
pyasn1==0.6.0
pycodestyle==2.11.1
pycparser==2.22
pycryptodome==3.10.4
pydeck==0.8.1b0
pyflakes==3.2.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyyaml==6.0.1
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
requests-file==2.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.29
stack-data==0.6.3
streamlit-authenticator==0.3.2
streamlit==1.33.0
supertokens-python==0.18.7
tenacity==8.2.3
tldextract==3.1.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
watchdog==4.0.0
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4